requirements.txt
streamlit
pandas
numpy
yfinance
scikit-learn
matplotlib
plotly
fpdf
openai
openai>=1.0.0
python-dotenv

